Split jpscore_test.py
Driver code now split into individual files. All contained in 'driver'.
Test reside now directly in 'systemtest'.
