+++ systemtest/driver/fixtures.py
@@ -1,0 +1,7 @@
+import pytest
+from driver.environment import Environment
+
+
+@pytest.fixture
+def env():
+    return Environment()
