+++ systemtest/driver/utils.py
@@ -1,37 +1,74 @@
 import pathlib
-import re
+import os
+import shutil
+import difflib
 
-import pandas as pd
+from driver.environment import Environment
+from driver.driver import JpsCoreDriver
 
 
-def read_traj_file(traj_file: pathlib.Path):
+def setup_jpscore_driver(
+    *,
+    env: Environment,
+    working_directory: pathlib.Path,
+    test_directory: pathlib.Path,
+):
     """
-    Reads in data from trajectory file.
+    Sets up jpscore driver by preparing the working directory and creating a driver.
 
-    :param traj_file (Pathlib.path): trajectory file to read in
-    :return fps (float), N (int), data (numpy array): frames per second, number of agents, trajectory data
-
+    :param env: global environment object
+    :param working_directory: working directory where jpscore should be executed
+    :param test_directory: directory of the test
+    :return: JpsCoreDriver object
     """
-    with open(traj_file, "r") as file:
-        for line in file:
-            if "#agents" in line:
-                # get number of agents
-                N = int(re.search(r"\d+", line).group())
-            if "#framerate" in line:
-                # get fps
-                fps = float(re.search(r"\d+.\d+", line).group())
-            if "#ID\tFR" in line:
-                # get column names
-                header = line.rstrip().strip("#").split("\t")
-                header.append("dummy after last tab")
-                break
+    copy_all_files(
+        src=env.systemtest_path / test_directory / "input",
+        dest=working_directory,
+    )
 
-    data = pd.read_table(
-        traj_file,
-        comment="#",
-        skip_blank_lines=True,
-        names=header,
-        usecols=["ID", "FR", "X", "Y", "Z"],
-    ).to_numpy()
+    jpscore_driver = JpsCoreDriver(
+        jpscore_path=env.jpscore_path, working_directory=working_directory
+    )
+    return jpscore_driver
 
-    return fps, N, data
+
+def get_file_text_diff(*, expected: pathlib.Path, actual: pathlib.Path):
+    """
+    Compares the actual with an expected text file and returns their difference.
+
+    :param expected: text file to expect
+    :param actual: actual text file
+    :return: string with differing text
+    """
+    with open(expected) as expected_data:
+        expected_text = expected_data.readlines()
+    with open(actual) as actual_data:
+        actual_text = actual_data.readlines()
+
+    diff_text = "".join(
+        difflib.context_diff(
+            expected_text,
+            actual_text,
+            fromfile=str(expected),
+            tofile=str(actual),
+        )
+    )
+
+    return diff_text
+
+
+def copy_all_files(*, src: pathlib.Path, dest: pathlib.Path):
+    """
+    Copies all files that exist in source directory to destination directory.
+    No directories are copied.
+
+    :param src: source directory
+    :param dest: destination directory
+    """
+    if src == dest:
+        return
+
+    for filename in os.listdir(src):
+        pathname = src / filename
+        if os.path.isfile(pathname):
+            shutil.copy2(pathname, dest)
